Extract writers from middleware.py
