+++ wsgi_lineprof/middleware.py
@@ -1,45 +1,10 @@
 from io import TextIOWrapper  # noqa: F401
-from queue import Queue
 import sys
-from threading import Thread
 from typing import Any, Callable, Iterable, TextIO, Union  # noqa: F401
 
 from wsgi_lineprof.profiler import LineProfiler
-from wsgi_lineprof.stats import FilterType, LineProfilerStats  # noqa: F401
-
-
-class SyncWriter(object):
-    def __init__(self,
-                 stream,  # type: Union[TextIO, TextIOWrapper]
-                 ):
-        # type: (...) -> None
-        self.stream = stream
-
-    def write(self, stats):
-        # type: (LineProfilerStats) -> None
-        stats.write_text(self.stream)
-
-
-class AsyncWriter(object):
-    def __init__(self,
-                 stream,  # type: Union[TextIO, TextIOWrapper]
-                 ):
-        # type: (...) -> None
-        self.stream = stream
-        self.queue = Queue()  # type: Queue
-        self.writer_thread = Thread(target=self._write)
-        self.writer_thread.setDaemon(True)
-        self.writer_thread.start()
-
-    def write(self, stats):
-        # type: (LineProfilerStats) -> None
-        self.queue.put(stats)
-
-    def _write(self):
-        # type: () -> None
-        while True:
-            stats = self.queue.get()
-            stats.write_text(self.stream)
+from wsgi_lineprof.stats import FilterType  # noqa: F401
+from wsgi_lineprof.writers import AsyncWriter, SyncWriter
 
 
 class LineProfilerMiddleware(object):
